refactor: tidy up CSV processing by separately organising preprocessing and schema validation
